+++ src/main/java/mate/academy/springboot/datajpa/service/ProductService.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import mate.academy.springboot.datajpa.model.Product;
 
@@ -10,7 +11,7 @@ public interface ProductService {
 
     void delete(Product product);
 
-    List<Product> findProductsByPriceBetween(Long from, Long to);
+    List<Product> findProductsByPriceBetween(BigDecimal from, BigDecimal to);
 
     List<Product> findProductsByCategoryName(List<String> categories);
 }
